Implement .hooks service method
